Test in ActiveRecord 4.2 and Ruby >=2.4
This increases the testing matrix by testing a newer ActiveRecord and a
newer Ruby.

ActiveRecord 3.2 does not work with Ruby 2.4 and above.

The ActiveRecord API changes between 3.2 and 4.2, which means that
`.all` and `.count` are no longer supported. The change is to use
`#where` instead.

The generated SQL changes slightly as well. Instead of putting the
parameterized values in the SQL string, a question mark is now outputted
instead. We use a version check to know which output we are testing for.