Remove upper boundary on ActiveRecord, test main
In order to keep this gem up to date with Rails main we'd like to start
testing against it on a schedule. Right now this will test against Rails
main daily. Once we confirm this gem is compatible with the current
Rails main we'll set it up to notify us if the main testing workflow
fails 3 days in a row so we can go and fix it. We decided on 3 days in a
row and not on _every_ failure just in case it only fails for some flaky reason.

I added `/gemfiles/activerecord_main.gemfile.lock` to `.gitignore`
because, for obvious reasons, we never want to lock the sha we're using.